Spawn the telemetry task only if needed
The telemetry tasks were all spawned together even if not needed and
kept idle if they were disabled by the configuration.
Change the behavior to spawn the tasks only if they are enabled and with
a period greater than 0.
Decouple the data retrieval from the send to Astarte using channels.

Signed-off-by: Francesco Vaiani <francesco.vaiani@secomind.com>