fix(test): racy assertion in test update after cancel
The test update after cancel would fail since it asserted that two to
download by the server. The two calls would only happen in a specific
sequence of messages exchanged by the channels between the `ota` and
`ota_handle` tasks. Sometimes the cancel message arrived before the file
was downloaded, and the assertion of two hits on the mock server would
fail. The change takes in consideration this by or-ing the two cases, it
also improves the tempdir creation by providing a prefix for each test.

Signed-off-by: Joshua Chapman <joshua.chapman@secomind.com>