feat(volume): inspect volume by name
Check if a volume with the same name already exists. This will prevent
craches/errors when the volume was already created, but the state was
not updated. We are using a unique id for the name of the volume and the
volume resource is immutable, so it's safe to assume a volume with the
same name is the same volume.

Signed-off-by: Joshua Chapman <joshua.chapman@secomind.com>